# contracts/http/swagger.yaml
basePath: /
definitions:
  apierror.Code:
    enum:
    - validation_failed
    - authentication_required
    - reauthentication_required
    - authorization_denied
    - not_found
    - conflict
    - rate_limited
    - csrf_rejected
    - origin_rejected
    - body_too_large
    - dependency_unavailable
    - internal_error
    type: string
    x-enum-varnames:
    - CodeValidation
    - CodeAuthentication
    - CodeReauthenticate
    - CodeAuthorization
    - CodeNotFound
    - CodeConflict
    - CodeRateLimited
    - CodeCSRF
    - CodeOrigin
    - CodeBodyTooLarge
    - CodeDependency
    - CodeInternal
  apierror.Detail:
    properties:
      code:
        $ref: '#/definitions/apierror.Code'
      correlation_id:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
  apierror.Response:
    properties:
      error:
        $ref: '#/definitions/apierror.Detail'
    type: object
...
info:
  contact: {}
  description: HTTP boundary exposed by the Quack v5 backend.
  title: Quack HTTP API
  version: "5.0"
paths:
  /auth/discord/callback:
    get:
      parameters:
      - description: Discord authorization code
        in: query
        name: code
        type: string
      - description: Single-use OAuth state
        in: query
        name: state
        required: true
        type: string
      - description: Discord OAuth error
        in: query
        name: error
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "302":
          description: Found
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apierror.Response'
      summary: Complete Discord sign-in
      tags:
      - Authentication
  /auth/discord/login:
    get:
      parameters:
      - description: Response mode
        enum:
        - redirect
        - json
        in: query
        name: mode
        type: string
      - description: Post-login relative redirect
        in: query
        name: redirect_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "302":
          description: Found
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apierror.Response'
      summary: Start Discord sign-in
      tags:
      - Authentication
  /auth/logout:
    post:
      responses:
        "204":
          description: No Content
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apierror.Response'
      security:
      - CookieAuth: []
      summary: Log out the current session
      tags:
      - Authentication
  /auth/logout-all:
    post:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apierror.Response'
      security:
      - CookieAuth: []
      summary: Log out every user session
      tags:
      - Authentication
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.Response'
      security:
      - CookieAuth: []
      summary: Get the current session
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
  MetricsKey:
    in: header
    name: X-Quack-Metrics-Key
    type: apiKey
  OpsKey:
    in: header
    name: X-Quack-Ops-Key
    type: apiKey
swagger: "2.0"